apps/personalai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
base58==2.1.1
imagehash==4.3.1
qrcode[pil]==7.4.2
eth-keys==0.4.0
eth-utils==2.3.1
nacl==1.5.0
# Music Generation (MusicGen from AudioCraft)
torch>=2.0.0
torchaudio>=2.0.0
audiocraft>=1.3.0
scipy>=1.10.0
soundfile>=0.12.0
librosa>=0.10.0
mido>=2.0.0
python-rtmidi>=2.3.0

